requirements.txt
pipecat-ai
pipecat-ai[sarvam]
pipecat-ai[google]
torch
fastapi
uvicorn
python-dotenv
loguru
aiohttp